Added passong tests for the new data processing pipeline
