+++ src/main/java/com/chilborne/covidradar/data/collection/DailyRecordDataCollector.java
@@ -1,0 +1,11 @@
+package com.chilborne.covidradar.data.collection;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public interface DailyRecordDataCollector<T> {
+
+    void collectData();
+
+    void publish(String data);
+}
