+++ src/main/java/com/chilborne/covidradar/data/processing/pipeline/DailyRecordDataPipeline.java
@@ -1,0 +1,24 @@
+package com.chilborne.covidradar.data.processing.pipeline;
+
+import com.chilborne.covidradar.events.NewDataEvent;
+import org.springframework.context.event.EventListener;
+import org.springframework.stereotype.Service;
+
+@Service
+public class DailyRecordDataPipeline {
+
+    private final Pipeline pipeline;
+
+    public DailyRecordDataPipeline(Pipeline pipeline) {
+        this.pipeline = pipeline;
+    }
+
+    @EventListener
+    public void startPipeline(NewDataEvent newDataEvent) {
+        pipeline.execute(newDataEvent.getData());
+    }
+
+
+
+
+}
