+++ src/main/java/com/chilborne/covidradar/data/processing/steps/Step.java
@@ -1,0 +1,15 @@
+package com.chilborne.covidradar.data.processing.steps;
+
+public interface Step<I, O> {
+    public static class PipeLineProcessException extends RuntimeException {
+        public PipeLineProcessException(Throwable t) {
+            super(t);
+        }
+
+        public PipeLineProcessException() {
+        }
+    }
+
+    public O process(I input) throws PipeLineProcessException;
+
+}
