--- src/main/java/com/chilborne/covidradar/services/dailyrecords/data/collection/DailyRecordDataCollector.java
@@ -1,11 +1,0 @@
-package com.chilborne.covidradar.services.dailyrecords.data.collection;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface DailyRecordDataCollector<T> {
-
-    void collectData();
-
-    void publish(String data);
-}
