--- src/main/java/com/chilborne/covidradar/services/dailyrecords/data/processing/steps/StaticDataVerifier.java
@@ -1,15 +1,0 @@
-package com.chilborne.covidradar.services.dailyrecords.data.processing.steps;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class StaticDataVerifier implements Step<String, String> {
-
-    @Override
-    public String process(String input) {
-
-        return input;
-    }
-
-
-}
